requirements.txt
crewai
crewai-tools
langchain-openai
langchain-community
pypdf
python-dotenv
spacy
sumy
fastapi
uvicorn[standard]
streamlit
pymongo
sentence-transformers
chromadb
faiss-cpu
openai
pydantic
pydantic-settings
httpx
requests
passlib[bcrypt]
pyjwt
cryptography
PyPDF2
motor
email-validator
python-multipart
matplotlib
pandas
seaborn